feat: Introduce Claude SDK modularization with unified import shim
- Added `sdk_import.py` to handle imports of `claude_agent_sdk` and `claude_code_sdk`, providing a single point of access and availability checks.
- Implemented `session.py` for session management, including `StepSessionClient` and `StepSession` classes to orchestrate work, finalize, and route phases.
- Created `session_types.py` to define data structures for session results, including `WorkPhaseResult`, `FinalizePhaseResult`, and `RoutePhaseResult`.
- Developed `telemetry.py` for tracking session metrics and tool call timings.
- Added contract tests to enforce import restrictions and verify public API stability for the new modular structure.
