+++ swarm/runtime/_claude_sdk/__init__.py
@@ -1,0 +1,4 @@
+"""Internal Claude SDK package.
+
+This package contains modularized components from claude_sdk.py.
+"""
